Extract sandbox runtime crate
